more utf8 fixes, font rendering fixes, auto download steam workshops, misc fixes

+++ code/cgame/cg_drawdc.h
@@ -8,14 +8,14 @@
 void CG_DrawHandlePicDc (float x, float y, float w, float h, qhandle_t asset, int widescreen, rectDef_t menuRect);
 void CG_DrawStretchPicDc (float x, float y, float w, float h, float s1, float t1, float s2, float t2, qhandle_t hShader, int widescreen, rectDef_t menuRect);
 void CG_DrawTextDc (float x, float y, float scale, const vec4_t color, const char *text, float adjust, int limit, int style, int fontIndex, int widescreen, rectDef_t menuRect);
-int CG_TextWidthDc (const char *text, float scale, int limit, int fontIndex, int widescreen, rectDef_t menuRect);
-int CG_TextHeightDc (const char *text, float scale, int limit, int fontIndex, int widescreen, rectDef_t menuRect);
+float CG_TextWidthDc (const char *text, float scale, int limit, int fontIndex, int widescreen, rectDef_t menuRect);
+float CG_TextHeightDc (const char *text, float scale, int limit, int fontIndex, int widescreen, rectDef_t menuRect);
 void CG_FillRectDc (float x, float y, float w, float h, const vec4_t color, int widescreen, rectDef_t menuRect);
 void CG_DrawRectDc (float x, float y, float w, float h, float size, const vec4_t color, int widescreen, rectDef_t menuRect);
 void CG_DrawSidesDc (float x, float y, float w, float h, float size, int widescreen, rectDef_t menuRect);
 void CG_DrawTopBottomDc (float x, float y, float w, float h, float size, int widescreen, rectDef_t menuRect);
 void CG_DrawTextWithCursorDc (float x, float y, float scale, const vec4_t color, const char *text, int cursorPos, char cursor, int limit, int style, int fontIndex, int widescreen, rectDef_t menuRect);
-int CG_OwnerDrawWidthDc (int ownerDraw, float scale, int fontIndex, int widescreen, rectDef_t menuRect);
+float CG_OwnerDrawWidthDc (int ownerDraw, float scale, int fontIndex, int widescreen, rectDef_t menuRect);
 
 int CG_PlayCinematicDc (const char *name, float x, float y, float w, float h, int widescreen, rectDef_t menuRect);
 void CG_StopCinematicDc (int handle);
